requirements.txt
aiogram==3.3.0
aiohttp==3.9.3
python-dotenv==1.0.1
asyncio==3.4.3 